Weather-App
This is a weather app created by me as a mini project during summer vacation 2021.This is based on node js http module to create a server and a request module to make HTTP call to an openweather.com API. Beside this, HTML,CSS,JavaScript were also used for the front end UI of the App.

Technology used- HTML,CSS,JavaScript,Node js.